FTS-122: User's subject directly for proxy issuer
Before we were using the userDn as passed from the application, but
then the canonical representation of the proxy's issuer and the user's certificate subject
would not match when using canonical representation.
This caused voms-clients 3 to fail complaining about different chains in the generated
proxy when getting the VO extensions.